skillsakhi_backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
